rename file name scroll button
